Add decoder wrapper to inference
